+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Bug Fix: quote query segments in generated types. See [#49](https://github.com/tatethurston/nextjs-routes/issues/49) for more context.
 - Bug Fix: don't generate routes for non navigable routes (`_error`, `_app`, `_document`).
+- Bug Fix: don't generate routes for test files that are co-located in pages directory. See [#50](https://github.com/tatethurston/nextjs-routes/pull/50) for more context.
 
 ## 0.0.18